Bound for test of weak keys
Added fn test_key(). This function can tests if your public key comes from a weak private key.
The inputs for this function are the public key and the bound. The bound is a usize: 32,64,128,160.
For example, if you use a bound 32, then you are testing your key looking in the soubgroup od Zp* of order less than 2^32.